feat: new chunk generation algorithm

+++ Configuration.cs
@@ -9,7 +9,7 @@ public partial class Configuration : Node
 	public static readonly float MOVEMENT_SPEED = 10;
 	public static readonly float GRAVITY = 20;
 	public static readonly float JUMP_VELOCITY = 10;
-	public static readonly float CHUNK_LOAD_RADIUS = 5;
+	public static readonly int CHUNK_LOAD_RADIUS = 5;
 	public static readonly float CHUNK_UNLOAD_TIME = 10;
 	public static readonly bool FLYING_ENABLED = true;
 }
